feat: Add `autoResize` for `SpriteGroupComponent` (#2442)
This PR adds an autoResize parameter to SpriteGroupComponent which will allow Flame users to control if the size of the component should be re-calculated to fit the size of current active sprite in the group. This new parameter auto-populates when not specified, based value of size while construction. If size is not given, autoResize will be true and vice-versa.

Caveat:
If someone goes and replaces the actual Sprite at a particular key within SpriteGroupComponent.sprites map and if that key is the current state, it will not trigger an auto-resize. This seems like a very unlikely use-case, so I am not worried about it too much.