feat: Add `autoResize` for `SpriteAnimationComponent` and `SpriteAnimationGroupComponent` (#2453)
This PR extends autoResize to SpriteAnimationComponent and SpriteAnimationGroupComponent similar to what was done in #2430 and #2442.

Additionally, this PR also adds playing property to SpriteAnimationGroupComponent. If it is desired to make this appear as a separate entry in the changlog, I can move this change to a separate PR as it is completely unrelated to the auto-resize changes.