docs: Fix mermaid graphs
Styling of the graphs were not proper and custom stylings were being
overrided by the default styling. This commit fixes that issue. There is
another issue where the arrow is not shown properly and a dimmed line is
shown, this is because of improper syntax, that also has ben changed in
this commit. And lastly, the component lifecycle graph onLoad is fixed.