docs: Fix mermaid graphs (#2483)
Styling of the graphs were not proper and custom stylings were being overrided by the default styling. This PR fixes that issue. There is another issue where the arrow is not shown properly and a dimmed line is shown, this is because of improper syntax, that also has ben changed in this commit. And lastly, the component lifecycle graph onLoad and low-level game api graph are fixed.

The yellow styling is defined before default styling which ignores yellow styling. So, moving it after default will work as expected.

Also, the syntax A-- This is a text .->B is not a syntax. So, changing it to either -- or -. | .- completely will fix the non arrow showing line.

This PR changes 3 graphs.