feat: Add next frame button to `FlameStudio` (#2522)
This PR adds a 3rd toolbar button to FlameStudio for stepping the engine frame by frame. This button will be active only when engine is paused. It steps with the default step time for now, but if/when FlameStudio gets user tweakable preferences, we can open up an option to allow users to override the step time.