fix: Animation ticker readability improvements (#2578)
This PR marks SpriteAnimation.ticker() as deprecated with SpriteAnimation.createTicker() as a replacement. This is done to make it more clear that calling that method creates a new ticker. Also, the docs for SpriteAnimationGroupComponent are updated with code snippet showing how to access and use the callbacks from tickers of each animation state.