feat: Expose atlas limits for `TiledComponent` (#2701)
This PR add the ability for users to override the max size of
`TiledAtlas` used by `TiledComponent` for packing all the images of all
the tilesets used by a tilemap. The overrides can be used if one is sure
that their target platform can handle huge textures.

It also adds an assert for cases when the `RectangleBinPacker` runs out
of space. This is done to help users find the problem quickly.