feat(flame_jenny): Allow removal of functions and commands (#2717)
Allow functions and commands to be removed from storage, or function and/or command storage to be cleared.

This is useful if different parts of the game have different implementations of a command or a function. Currently registering a function or command with the same name fails (i.e. it is not overwritten). Removing the function or command allows it to be replaced.