docs: Replace flutter format with dart format (#2725)
flutter format will generate a deprecation warning:

The "format" command is deprecated. Please use the "dart format" sub-command instead, which has the same command-line usage as "flutter format".
It should be replaced with dart format.

Have updated two places in the docs, and in a shell script (please verify this one for me!)