feat: Implement Snapshot mixin on PositionComponent (#2695)
This implements a Snapshot mixin on PositionComponent that allows:

    A component to be rendered once then cached (for performance)
    and/or a Picture or Image snapshot to be taken at any time

Internally the snapshot is cached and redrawn as a Picture. After testing with both Image and Picture I found that Image had no performance improvement, and the rendering quality was lower.

A helper method snapshotToImage was added to allow the user better control of generating an Image from the internal Picture snapshot if that is desired.

See the documentation in this PR for more details.