fix: Correct FlameGame lifecycle diagram and description (#2707)
The FlameGame lifecycle diagram is incorrect. The lifecycle for a FlameGame is as follows:

onGameResize
onLoad
onMount
render
update
onRemove
onGameResize
onLoad
onMount
render
update
Note in particular the onGameResize is before onLoad and the render and update are reversed. Because they are different, this PR creates a new lifecycle diagram that is not shared with Component. It also fixes the Component lifecycle diagram to be consistent with the FlameGame lifecycle diagram formatting and terms.