fix: Parallax offset calculations in flame_tiled don't scale properly (#2766)
The new camera systems seems to have uncovered a bug in the flame_tiled parallax calculations. The camera zoom levels need to be scaled by the tile size to by applied in the correct coordinate space. I've mostly validated this manually by comparing the Tiled application's viewport to a Flame game at runtime.